created shared.js, refactored to file
