+++ lib/index.js
@@ -1,3 +1,4 @@
+require('./shared')
 require('./foods')
-require('./foods.scss')
 require('./diary')
+require('./foods.scss')
